Data.Semigroup

Description

In mathematics, a semigroup is an algebraic structure consisting of a set together with an associative binary operation. A semigroup generalizes a monoid in that there might not exist an identity element. It also (originally) generalized a group (a monoid with all inverses) to a type where every element did not have to have an inverse, thus the name semigroup.

The use of (<>) in this module conflicts with an operator with the same name that is being exported by Data.Monoid. However, this package re-exports (most of) the contents of Data.Monoid, so to use semigroups and monoids in the same package just

import Data.Semigroup

Documentation

class Semigroup a where

Minimal complete definition

Nothing

Methods

(<>) :: a -> a -> a infixr 6

An associative operation.

 (a <> b) <> c = a <> (b <> c)
 

If a is also a Monoid we further require

 (<>) = mappend
 

sconcat :: NonEmpty a -> a

Reduce a non-empty list with <>

The default definition should be sufficient, but this can be overridden for efficiency.

times1p :: Whole n => n -> a -> a

Repeat a value (n + 1) times.

 times1p n a = a <> a <> ... <> a  -- using <> n times
 

The default definition uses peasant multiplication, exploiting associativity to only require O(log n) uses of <>.

See also timesN.

Semigroups

newtype Min a

Constructors

Min

Fields

getMin :: a

newtype Max a

Constructors

Max

Fields

getMax :: a

newtype First a

Use Option (First a) to get the behavior of First from Data.Monoid.

Constructors

First

Fields

getFirst :: a

newtype Last a

Use Option (Last a) to get the behavior of Last from Data.Monoid

Constructors

Last

Fields

getLast :: a

newtype WrappedMonoid m

Provide a Semigroup for an arbitrary Monoid.

Constructors

WrapMonoid

Fields

unwrapMonoid :: m

timesN :: (Whole n, Monoid a) => n -> a -> a

Repeat a value n times.

timesN n a = a <> a <> ... <> a  -- using <> (n-1) times

Implemented using times1p.

Re-exported monoids from Data.Monoid

class Monoid a where

Minimal complete definition

mempty, mappend

Methods

mempty :: a

mappend :: a -> a -> a

mconcat :: [a] -> a

newtype Dual a

Constructors

Dual

Fields

getDual :: a

newtype Endo a

Constructors

Endo

Fields

appEndo :: a -> a

newtype All

Constructors

All

Fields

getAll :: Bool

newtype Any

Constructors

Any

Fields

getAny :: Bool

newtype Sum a

Constructors

Sum

Fields

getSum :: a

newtype Product a

Constructors

Product

Fields

getProduct :: a

A better monoid for Maybe

newtype Option a

Option is effectively Maybe with a better instance of Monoid, built off of an underlying Semigroup instead of an underlying Monoid.

Ideally, this type would not exist at all and we would just fix the Monoid instance of Maybe

Constructors

Option

Fields

getOption :: Maybe a

option :: b -> (a -> b) -> Option a -> b

Fold an Option case-wise, just like maybe.

Difference lists of a semigroup

diff :: Semigroup m => m -> Endo m

This lets you use a difference list of a Semigroup as a Monoid.

cycle1 :: Semigroup m => m -> m

A generalization of cycle to an arbitrary Semigroup. May fail to terminate for some values in some semigroups.