Data.Semigroup.Generic

Description

This module provides generic deriving tools for monoids and semigroups for product-like structures.

Generic method implementations

gmappend :: (Generic a, GSemigroup (Rep a)) => a -> a -> a

Generically generate a Semigroup (<>) operation for any type implementing Generic. This operation will append two values by point-wise appending their component fields. It is only defined for product types.

gmappend a (gmappend b c) = gmappend (gmappend a b) c

gmempty :: (Generic a, GMonoid (Rep a)) => a

Generically generate a Monoid mempty for any product-like type implementing Generic.

It is only defined for product types.

gmappend gmempty a = a = gmappend a gmempty

Adapter newtype

newtype GenericSemigroupMonoid a

An adapter newtype, suitable for DerivingVia. Its Semigroup and Monoid instances leverage the Generic-based defaults defined by gmappend and gmempty. Here is an example of how to use it:

{-# LANGUAGE DerivingVia #-}
import Data.Semigroup.Generic

data Pair a = MkPair a a
  deriving (Semigroup, Monoid) via (GenericSemigroupMonoid (Pair a))

Constructors

GenericSemigroupMonoid

Fields

getGenericSemigroupMonoid :: a

Internal classes

class GSemigroup f

Minimal complete definition

gmappend'

class GSemigroup f => GMonoid f

Minimal complete definition

gmempty'