Data.Semigroup

Description

In mathematics, a semigroup is an algebraic structure consisting of a set together with an associative binary operation. A semigroup generalizes a monoid in that there might not exist an identity element. It also (originally) generalized a group (a monoid with all inverses) to a type where every element did not have to have an inverse, thus the name semigroup.

The use of (<>) in this module conflicts with an operator with the same name that is being exported by Data.Monoid. However, this package re-exports (most of) the contents of Data.Monoid, so to use semigroups and monoids in the same package just

 import Data.Semigroup

Documentation

class Semigroup a where

Methods

(<>) :: a -> a -> a

An associative operation.

 (a <> b) <> c = a <> (b <> c)

If a is also a Monoid we further require

 (<>) = mappend

sconcat :: NonEmpty a -> a

Reduce a non-empty list with <>

The default definition should be sufficient, but this can be overridden for efficiency.

times1p :: Whole n => n -> a -> a

Repeat a value (n + 1) times.

 times1p n a = a <> a <> ... <> a  -- using <> n times

The default definition uses peasant multiplication, exploiting associativity to only require O(log n) uses of <>.

See also times.

Semigroups

newtype Min a

Constructors

Min

Fields

getMin :: a

newtype Max a

Constructors

Max

Fields

getMax :: a

newtype First a

Use Option (First a) -- to get the behavior of First

Constructors

First

Fields

getFirst :: a

newtype Last a

Use Option (Last a) -- to get the behavior of Last

Constructors

Last

Fields

getLast :: a

newtype WrappedMonoid m

Provide a Semigroup for an arbitrary Monoid.

Constructors

WrapMonoid

Fields

unwrapMonoid :: m

timesN :: (Whole n, Monoid a) => n -> a -> a

Repeat a value n times.

 times n a = a <> a <> ... <> a  -- using <> (n-1) times

Implemented using times1p.

Re-exported monoids from Data.Monoid

class Monoid a where

The class of monoids (types with an associative binary operation that has an identity). Instances should satisfy the following laws:

mappend mempty x = x

mappend x mempty = x

mappend x (mappend y z) = mappend (mappend x y) z

mconcat = foldr mappend mempty

The method names refer to the monoid of lists under concatenation, but there are many other instances.

Minimal complete definition: mempty and mappend.

Some types can be viewed as a monoid in more than one way, e.g. both addition and multiplication on numbers. In such cases we often define newtypes and make those instances of Monoid, e.g. Sum and Product.

Methods

mempty :: a

Identity of mappend

mappend :: a -> a -> a

An associative operation

mconcat :: [a] -> a

Fold a list using the monoid. For most types, the default definition for mconcat will be used, but the function is included in the class definition so that an optimized version can be provided for specific types.

newtype Dual a

The dual of a monoid, obtained by swapping the arguments of mappend.

Constructors

Dual

Fields

getDual :: a

newtype Endo a

The monoid of endomorphisms under composition.

Constructors

Endo

Fields

appEndo :: a -> a

newtype All

Boolean monoid under conjunction.

Constructors

All

Fields

getAll :: Bool

newtype Any

Boolean monoid under disjunction.

Constructors

Any

Fields

getAny :: Bool

newtype Sum a

Monoid under addition.

Constructors

Sum

Fields

getSum :: a

newtype Product a

Monoid under multiplication.

Constructors

Product

Fields

getProduct :: a

A better monoid for Maybe

newtype Option a

Option is effectively Maybe with a better instance of Monoid, built off of an underlying Semigroup instead of an underlying Monoid. Ideally, this type would not exist at all and we would just fix the Monoid intance of Maybe

Constructors

Option

Fields

getOption :: Maybe a

option :: b -> (a -> b) -> Option a -> b

Difference lists of a semigroup

diff :: Semigroup m => m -> Endo m

This lets you use a difference list of a Semigroup as a Monoid.

cycle1 :: Semigroup m => m -> m

A generalization of cycle to an arbitrary Semigroup. May fail to terminate for some values in some semigroups.